requirements.txt
python-nmap
psutil
paramiko
colorama
ftplib